Renaming and further separation of each class for its own source file
