+++ src/Fluid1DLib.cpp
@@ -1,20 +1,13 @@
-#include "includes/Tethys1DLib.h"
+#include "includes/Fluid1DLib.h"
 #include "SetUpParametersLib.h"
+#include "GrapheneFluid1DLib.h"
 
 
 using namespace H5;
 using namespace std;
 
 
-GrapheneFluid1D::GrapheneFluid1D(SetUpParameters &input_parameters) : Fluid1D(input_parameters) {
-	vel_fer = input_parameters.FermiVelocity;
-	col_freq = input_parameters.CollisionFrequency;
-	char buffer [50];
-	sprintf (buffer, "S=%.2fvF=%.2fvis=%.2fl=%.2f", vel_snd, vel_fer, kin_vis, col_freq);
-	file_infix = buffer;
-}
-
-/*....................................................................*/	
+/*....................................................................*/
 /*.......... 1 Dimensional Fluid Class ...............................*/	
 /*....................................................................*/
 Fluid1D::Fluid1D(const SetUpParameters &input_parameters) : TethysBase{input_parameters.SizeX, 0, 1}{
@@ -38,20 +31,6 @@ Fluid1D::Fluid1D(const SetUpParameters &input_parameters) : TethysBase{input_par
 }	
 
 Fluid1D::~Fluid1D() = default;
-
-GrapheneFluid1D::~GrapheneFluid1D(){
-	delete Den;
-	delete Vel ;
-	delete Cur ;
-	delete den_mid ;
-	delete vel_mid ;
-	delete DenCor ;
-	delete VelCor ;
-	delete CurCor ;
-	delete vel_snd_arr ;
-	delete GradVel ;
-	delete grad_vel_mid ;
-}
 
 float  Fluid1D::DensityFlux(float n,float v, __attribute__((unused)) float s){
 	float f_1;
@@ -179,25 +158,7 @@ void Fluid1D::Richtmyer(){
 } 
 
 /*....................................................................*/	
-/*............ Derived Graphene Class  ...............................*/	
-/*....................................................................*/	
-
-float GrapheneFluid1D::DensityFlux(float n,float v,float __attribute__((unused)) s){
-	float f_1;
-	f_1 = n * v;
-	return f_1;
-}
-float GrapheneFluid1D::VelocityFlux(float n,float v,__attribute__((unused))float dv,float s){
-	float f_2;
-		f_2 = 0.25f * v * v + vel_fer * vel_fer * 0.5f * log(n) + 2.0f * s * s * sqrt(n);//- kin_vis * dv;
-	return f_2;
-}
-float GrapheneFluid1D::DensitySource(__attribute__((unused)) float n, __attribute__((unused)) float v, __attribute__((unused)) float s){
-	return 0.0f;
-}
-float GrapheneFluid1D::VelocitySource(__attribute__((unused)) float n,float v,__attribute__((unused)) float s){
-	return -1.0f * col_freq * v ;
-}
+/*............ Derived Graphene Class  ...............................*/
 
 
 int Fluid1D::GetSnapshotStep() const { return snapshot_step;}
@@ -245,13 +206,3 @@ void Fluid1D::SaveSnapShot(){
 
 
 
-void GrapheneFluid1D::CflCondition(){
-	dx = lengX / ( float ) ( Nx - 1 );
-	float lambda;
-	if(vel_snd<0.36f*vel_fer){
-		lambda=1.2f*vel_fer;
-	}else{
-		lambda=1.97f*vel_snd + 0.5f*vel_fer;
-	}
-	dt = dx/lambda;
-}
