+++ src/TETHYS_1D_ElectronicAnalysis.cpp
@@ -11,7 +11,7 @@
 #include <numeric>
 //#include <fftw3.h>
 
-#include "includes/Tethys1DLib.h"
+#include "includes/Fluid1DLib.h"
 
 using namespace std;
