+++ src/TETHYS_1D_Main_v134.cpp
@@ -1,8 +1,9 @@
-#include "includes/Tethys1DLib.h"
+#include "includes/Fluid1DLib.h"
 #include "includes/BoundaryLib.h"
 #include "includes/ElectricLib.h"
 #include "SetUpParametersLib.h"
 #include "DyakonovShurBoundaryLib.h"
+#include "GrapheneFluid1DLib.h"
 
 #ifndef MAT_PI
 #	define MAT_PI 3.14159265358979323846
