+++ src/TETHYS_2D_ElectronicAnalysis.cpp
@@ -1,6 +1,7 @@
-#include "includes/Tethys2DLib.h"
+#include "includes/Fluid2DLib.h"
 #include "includes/ElectricLib.h"
 #include "SetUpParametersLib.h"
+#include "GrapheneFluid2DLib.h"
 
 
 using namespace std;
