+++ src/includes/BoundaryLib.h
@@ -9,8 +9,8 @@
 #include <H5Cpp.h>
 #include "TethysBaseLib.h"
 #include "TethysMathLib.h"
-#include "Tethys1DLib.h"
-#include "Tethys2DLib.h"
+#include "Fluid1DLib.h"
+#include "Fluid2DLib.h"
 
 
 //TODO review the entire boundary class
