+++ src/includes/Fluid1DLib.h
@@ -3,8 +3,8 @@
 * Distributed under the MIT License (license terms are at http://opensource.org/licenses/MIT).   *
 \************************************************************************************************/
 
-#ifndef TETHYS1DLIB_H
-#define TETHYS1DLIB_H
+#ifndef FLUID1DLIB_H
+#define FLUID1DLIB_H
 
 #include <H5Cpp.h>
 #include "TethysBaseLib.h"
@@ -60,23 +60,5 @@ class Fluid1D : public TethysBase{
 		int GetSnapshotFreq() const;
 };
 
-/*!
- * @brief Graphene electronic fluid class in one dimension.
- *
- * The GrapheneFluid1D class describes a  fluid governed by the usual continuity and Cauchy momemtum equations, where the mass of the fluid element is *not* constant.
- * It overrides class Fluid1D necessary methods in order to describe the semi-classical electronic fluid.
- * */
-class GrapheneFluid1D : public Fluid1D{
-	public :
-		explicit GrapheneFluid1D(SetUpParameters &input_parameters);
-		~GrapheneFluid1D();
-		/*Override CFL condition to the case of graphene equations */
-		void CflCondition() override;
-		/*Override fluxes and sources to specifics of graphene physics*/
-		float DensityFlux(float n,float v,__attribute__((unused)) float s) override;
-		float VelocityFlux(float n,float v,float dv,float s) override;
-		float DensitySource(__attribute__((unused)) float n,__attribute__((unused)) float v, __attribute__((unused)) float s) override;
-		float VelocitySource(__attribute__((unused)) float n,float v,__attribute__((unused)) float s) override;
-};
 #endif
